Fixed migration and added fields for the model

+++ database/migrations/2021_05_20_173135_create_sms_codes_table.php
@@ -15,7 +15,7 @@ class CreateSmsCodesTable extends Migration
     {
         Schema::create('sms_codes', function (Blueprint $table) {
             $table->id();
-            $table->unsignedTinyInteger('code');
+            $table->smallInteger('code');
             $table->timestamp('created_at');
             $table->timestamp('valid_time');
             $table->timestamp('timeout');
